% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{Imputation}
\alias{Imputation}
\title{\code{Imputation}}
\usage{
Imputation(data = NULL, formula = NULL, method = "try mice", m = 1,
  seed = 12321)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}

\item{formula}{A \code{\link{formula}}. Where the formula contains a dependent variable,
observations with missing values on this variable are deleted after the imputation (von Hippel 2007).}

\item{method}{"mice" applies multivariate imputation by chained equations
(predictive mean matching) with the \code{\link[mice]{mice}} package. "hot deck" applies the \code{\link[hot.deck:hot.deck]{hot.deck}} method.
The default setting is "try mice", which first applies the \code{\link[mice]{mice}} method and,if an error
occurs, falls back to \code{\link[hot.deck]{hot.deck}}.}

\item{m}{Number of imputation samples.}

\item{seed}{Seed used in random number generation.}
}
\description{
This is a wrapper for \code{\link{mice}}.
}
\references{
von Hippel, Paul T. 2007. "Regression With Missing Y's: An
Improved Strategy for Analyzing Multiply Imputed Data." Sociological Methodology 37:83-117.
Skyler J. Cranmer and Jeff Gill (2013). We Have to Be Discrete About This: A Non-Parametric Imputation Technique for Missing Categorical Data. British Journal of Political Science, 43, pp 425-449.
Stef van Buuren and Karin Groothuis-Oudshoorn (2011), "mice: Multivariate
Imputation by Chained Equations in R", Journal of Statistical Software, 45:3, 1-67.
}
